Fix a few other format strings.

+++ src/task.c
@@ -758,7 +758,8 @@ DLLEXPORT void gdblookup(ptrint_t ip)
         else if (line_num == -1)
             jl_safe_printf("%s at %s (unknown line)\n", func_name, file_name);
         else
-            jl_safe_printf("%s at %s:%d\n", func_name, file_name, line_num);
+            jl_safe_printf("%s at %s:%u\n", func_name, file_name,
+                           (unsigned)line_num);
     }
 }
